requirements.txt
# Core dependencies for epub2audiobook pipeline
numpy>=1.21.0
PyYAML>=6.0
tqdm>=4.64.0
soundfile>=0.12.1

# Optional TTS engine dependencies (uncomment as needed)
# TTS>=0.20.0                    # Coqui TTS for XTTS models
# torch>=1.13.0                 # PyTorch for neural TTS models
# transformers>=4.20.0          # Hugging Face models
# piper-tts>=1.2.0             # Piper fast TTS
# espeak-ng                    # eSpeak NG for phoneme generation

# Audio processing (for advanced features)
# librosa>=0.9.0               # Audio analysis
# scipy>=1.9.0                 # Signal processing
# noisereduce>=2.0.0           # Noise reduction

# Development and testing
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0